+++ app/src/main/java/com/smashing/app/data/mapper/SignUpMapper.kt
@@ -1,7 +1,7 @@
 package com.smashing.app.data.mapper
 
 import com.smashing.app.data.model.auth.SignUpModel
-import com.smashing.app.data.remote.dto.PostSignUpResponse
+import com.smashing.app.data.remote.dto.auth.PostSignUpResponse
 
 fun PostSignUpResponse.toSignUpModel() = SignUpModel(
     accessToken = this.accessToken,
