+++ app/src/main/java/com/smashing/app/data/remote/datasource/api/KakaoRegionDataSource.kt
@@ -1,6 +1,6 @@
 package com.smashing.app.data.remote.datasource.api
 
-import com.smashing.app.data.remote.dto.region.kakaoRegion.GetKakaoAddressSearchResponse
+import com.smashing.app.data.remote.dto.region.GetKakaoAddressSearchResponse
 
 interface KakaoRegionDataSource {
     suspend fun searchAddress(query: String): GetKakaoAddressSearchResponse
