+++ app/src/main/java/com/smashing/app/data/remote/datasource/impl/AuthRemoteDataSourceImpl.kt
@@ -2,10 +2,10 @@ package com.smashing.app.data.remote.datasource.impl
 
 import com.smashing.app.data.remote.datasource.api.AuthRemoteDataSource
 import com.smashing.app.data.remote.dto.BaseResponse
-import com.smashing.app.data.remote.dto.PostKakaoLoginRequest
-import com.smashing.app.data.remote.dto.PostKakaoLoginResponse
-import com.smashing.app.data.remote.dto.PostSignUpRequest
-import com.smashing.app.data.remote.dto.PostSignUpResponse
+import com.smashing.app.data.remote.dto.auth.PostKakaoLoginRequest
+import com.smashing.app.data.remote.dto.auth.PostKakaoLoginResponse
+import com.smashing.app.data.remote.dto.auth.PostSignUpRequest
+import com.smashing.app.data.remote.dto.auth.PostSignUpResponse
 import com.smashing.app.data.remote.service.AuthService
 import javax.inject.Inject
