+++ app/src/main/java/com/smashing/app/data/remote/dto/auth/PostSignUpRequest.kt
@@ -1,4 +1,4 @@
-package com.smashing.app.data.remote.dto
+package com.smashing.app.data.remote.dto.auth
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
