+++ app/src/main/java/com/smashing/app/data/repository/api/AuthRepository.kt
@@ -3,7 +3,7 @@ package com.smashing.app.data.repository.api
 import android.content.Context
 import com.smashing.app.data.model.auth.KakaoLoginModel
 import com.smashing.app.data.model.auth.SignUpModel
-import com.smashing.app.data.remote.dto.PostSignUpRequest
+import com.smashing.app.data.remote.dto.auth.PostSignUpRequest
 
 interface AuthRepository {
     suspend fun loginKakao(context: Context): Result<String>
