+++ app/src/main/java/com/smashing/app/data/repository/api/RegionRepository.kt
@@ -1,6 +1,6 @@
 package com.smashing.app.data.repository.api
 
-import com.smashing.app.data.model.KakaoRegion
+import com.smashing.app.data.model.region.KakaoRegion
 
 interface RegionRepository {
     suspend fun searchAddress(query: String): Result<List<KakaoRegion>>
