+++ app/src/main/java/com/smashing/app/data/repository/impl/RegionRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.smashing.app.data.repository.impl
 
 import com.smashing.app.core.util.suspendRunCatching
 import com.smashing.app.data.mapper.toRegionList
-import com.smashing.app.data.model.KakaoRegion
+import com.smashing.app.data.model.region.KakaoRegion
 import com.smashing.app.data.remote.datasource.api.KakaoRegionDataSource
 import com.smashing.app.data.repository.api.RegionRepository
 import javax.inject.Inject
